Move impersonator to root level query and resolve other remarks
