+++ packages/api/src/schema/types/user.ts
@@ -153,7 +153,6 @@ export default [
       allowedActions: [RoleAction!]!
       tokens: [Token!]!
       oAuthAccounts: [OAuthAccount!]!
-      impersonator: User
       reviews(limit: Int = 10, offset: Int = 0, sort: [SortOptionInput!]): [ProductReview!]!
       reviewsCount: Int!
     }
